+++ Assets/Editor/Adaptive Performance/Settings/Google Android Provider Settings.asset
@@ -12,12 +12,12 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 19bedd638b17440c6934f09de14d81c5, type: 3}
   m_Name: Google Android Provider Settings
   m_EditorClassIdentifier: 
-  m_Logging: 1
+  m_Logging: 0
   m_AutomaticPerformanceModeEnabled: 1
   m_EnableBoostOnStartup: 1
   m_StatsLoggingFrequencyInFrames: 50
   m_IndexerSettings:
-    m_Active: 1
+    m_Active: 0
     m_ThermalActionDelay: 10
     m_PerformanceActionDelay: 4
   m_ScalerSettings:
