+++ Assets/MyAsset/03. Script/Controller/InputObject/ClickableObject.cs
