+++ Assets/MyAsset/03. Script/Entity/BlockParticle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 17f9e15d4d908fe4fa31bd26d2b3c51b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
