+++ Assets/MyAsset/03. Script/Entity/Item/DroppedItem.meta
