+++ Assets/MyAsset/03. Script/Entity/Item/DroppedItem/DroppedEnergy.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace LaserCrush.Entity
+namespace LaserCrush.Entity.Item
 {
     public sealed class DroppedEnergy : DroppedItem
     {
