+++ Assets/MyAsset/03. Script/Entity/Item/DroppedItem/DroppedEnergy.cs.meta
