+++ Assets/MyAsset/03. Script/Entity/Item/DroppedItem/DroppedPrism.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace LaserCrush.Entity
+namespace LaserCrush.Entity.Item
 {
     public sealed class DroppedPrism : DroppedItem
     {
