+++ Assets/MyAsset/03. Script/Entity/Item/InstalledItem.meta
