+++ Assets/MyAsset/03. Script/Entity/Laser.cs
@@ -1,7 +1,8 @@
 using System.Collections.Generic;
-using UnityEngine;
 using System;
+using UnityEngine;
 using LaserCrush.Manager;
+using LaserCrush.Entity.Particle;
 
 namespace LaserCrush.Entity
 {
