+++ Assets/MyAsset/03. Script/Entity/LaserParticle.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace LaserCrush.Entity
+namespace LaserCrush.Entity.Particle
 {
     public class LaserParticle
     {
