+++ Assets/MyAsset/03. Script/Extension/ExtensionVector2.cs
@@ -21,7 +21,7 @@ namespace LaserCrush.Extension
         {
             float angle = Mathf.Atan2(direction.y, direction.x) * Mathf.Rad2Deg;
 
-            // x축 양 기준으로 각도를 15도에서 165도로 제한
+            // x축 양 기준으로 각도를 min도에서 max도로 제한
             float clampedAngle = Mathf.Clamp(angle, min, max);
 
             return new Vector2(Mathf.Cos(clampedAngle * Mathf.Deg2Rad), Mathf.Sin(clampedAngle * Mathf.Deg2Rad));
