+++ Assets/MyAsset/03. Script/UI/AcquiredItemUI.cs.meta
