+++ Assets/MyAsset/04. Prefab/Block/BlockParticle.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: b3da0e65a9871104ea505fbe58c741d6
-folderAsset: yes
-DefaultImporter:
+guid: 449ad90b4aa993a4fb34fcfa208f9910
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
