+++ Assets/MyAsset/04. Prefab/Item/InstalledItem/Prism2Branch.prefab
@@ -118,8 +118,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 180}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 0.4, y: 0.5}
-  m_SizeDelta: {x: 6.5, y: 7}
+  m_AnchoredPosition: {x: 1, y: 0.75}
+  m_SizeDelta: {x: 9.5, y: 10}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &951215897272433284
 CanvasRenderer:
@@ -149,7 +149,7 @@ MonoBehaviour:
   m_OnCullStateChanged:
     m_PersistentCalls:
       m_Calls: []
-  m_Sprite: {fileID: 21300000, guid: 2d93aa4a7897c4c4c8a87f22b40dcb4c, type: 3}
+  m_Sprite: {fileID: 21300000, guid: 0ddf1d90d7660494cadc9eb9d1a12c7d, type: 3}
   m_Type: 3
   m_PreserveAspect: 0
   m_FillCenter: 1
@@ -263,7 +263,7 @@ LineRenderer:
       m_RotationOrder: 4
     colorGradient:
       serializedVersion: 2
-      key0: {r: 0.40000004, g: 0.9960785, b: 0.9960785, a: 0.7058824}
+      key0: {r: 0.32941177, g: 0.6392157, b: 0.9725491, a: 0.7058824}
       key1: {r: 0.40000004, g: 0.9960785, b: 0.9960785, a: 0.23529412}
       key2: {r: 0, g: 0, b: 0, a: 0}
       key3: {r: 0, g: 0, b: 0, a: 0}
@@ -280,15 +280,15 @@ LineRenderer:
       ctime6: 0
       ctime7: 0
       atime0: 0
-      atime1: 36044
+      atime1: 26792
       atime2: 0
       atime3: 0
       atime4: 0
       atime5: 0
       atime6: 0
       atime7: 0
-      m_Mode: 0
-      m_ColorSpace: -1
+      m_Mode: 2
+      m_ColorSpace: 0
       m_NumColorKeys: 2
       m_NumAlphaKeys: 2
     numCornerVertices: 0
@@ -406,7 +406,7 @@ LineRenderer:
       m_RotationOrder: 4
     colorGradient:
       serializedVersion: 2
-      key0: {r: 0.40000004, g: 0.9960785, b: 0.9960785, a: 0.7058824}
+      key0: {r: 0.32941177, g: 0.6392157, b: 0.9725491, a: 0.7058824}
       key1: {r: 0.40000004, g: 0.9960785, b: 0.9960785, a: 0.23529412}
       key2: {r: 0, g: 0, b: 0, a: 0}
       key3: {r: 0, g: 0, b: 0, a: 0}
@@ -423,15 +423,15 @@ LineRenderer:
       ctime6: 0
       ctime7: 0
       atime0: 0
-      atime1: 36044
+      atime1: 26792
       atime2: 0
       atime3: 0
       atime4: 0
       atime5: 0
       atime6: 0
       atime7: 0
-      m_Mode: 0
-      m_ColorSpace: -1
+      m_Mode: 2
+      m_ColorSpace: 0
       m_NumColorKeys: 2
       m_NumAlphaKeys: 2
     numCornerVertices: 0
@@ -479,8 +479,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 180, z: 180}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: -0.4, y: 0.5}
-  m_SizeDelta: {x: 6.5, y: 7}
+  m_AnchoredPosition: {x: -1, y: 0.75}
+  m_SizeDelta: {x: 9.5, y: 10}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!222 &6495259297388507761
 CanvasRenderer:
@@ -510,7 +510,7 @@ MonoBehaviour:
   m_OnCullStateChanged:
     m_PersistentCalls:
       m_Calls: []
-  m_Sprite: {fileID: 21300000, guid: 2d93aa4a7897c4c4c8a87f22b40dcb4c, type: 3}
+  m_Sprite: {fileID: 21300000, guid: 0ddf1d90d7660494cadc9eb9d1a12c7d, type: 3}
   m_Type: 3
   m_PreserveAspect: 0
   m_FillCenter: 1
